pyproject.toml
[project]
name = "checklist"
version = "0.1.0"
description = "Telegram bot for computer club checklists"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.17.0",
    "aiohttp>=3.11.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.14.0",
    "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
